# src/main/java/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "CARGO MANAGEMENT SYSTEM.",
    "version" : "1.0.0",
    "title" : "Swagger Cargo",
    "termsOfService" : "http://swagger.io/terms/",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath" : "/v1",
  "tags" : [ {
    "name" : "customer",
    "description" : "Everything about customers",
    "externalDocs" : {
      "description" : "Find out more",
      "url" : "http://swagger.io"
    }
  }, {
    "name" : "receipt",
    "description" : "Everything about receipt: new, lookup..."
  }, {
    "name" : "user",
    "description" : "Operations about user"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Created user object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/user/display" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Logs user into the system",
        "description" : "",
        "operationId" : "getAllUsers",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            },
            "headers" : {
              "X-Rate-Limit" : {
                "type" : "integer",
                "format" : "int32",
                "description" : "calls per hour allowed by the user"
              },
              "X-Expires-After" : {
                "type" : "string",
                "format" : "date-time",
                "description" : "date in UTC when token expires"
              }
            }
          },
          "400" : {
            "description" : "Invalid username/password supplied"
          }
        }
      }
    },
    "/user/login" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Logs user into the system",
        "description" : "",
        "operationId" : "loginUser",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "The user name for login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "The password for login in clear text",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "headers" : {
              "X-Rate-Limit" : {
                "type" : "integer",
                "format" : "int32",
                "description" : "calls per hour allowed by the user"
              },
              "X-Expires-After" : {
                "type" : "string",
                "format" : "date-time",
                "description" : "date in UTC when token expires"
              }
            }
          },
          "400" : {
            "description" : "Invalid username/password supplied"
          }
        }
      }
    },
    "/user/logout" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Logs out current logged in user session",
        "description" : "",
        "operationId" : "logoutUser",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "Customer" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "fullAddress" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "emailAddress" : {
          "type" : "string"
        }
      }
    },
    "Sender" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "city" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string"
        },
        "zipcode" : {
          "type" : "string"
        }
      }
    },
    "Recipient" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ward" : {
          "type" : "string"
        },
        "district" : {
          "type" : "string"
        },
        "province" : {
          "type" : "string"
        }
      }
    },
    "ServiceType" : {
      "type" : "string",
      "enum" : [ "DOOR_TO_DOOR", "AIRPORT_TO_AIRPORT", "OTHER" ]
    },
    "Package" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int64"
        },
        "itemName" : {
          "type" : "string"
        }
      }
    },
    "TrackingNumber" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "description" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "roleId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "Role" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleName" : {
          "type" : "string"
        },
        "paymentMethodType" : {
          "$ref" : "#/definitions/PaymentMethod"
        },
        "createdDate" : {
          "type" : "string"
        }
      }
    },
    "PaymentMethod" : {
      "type" : "string",
      "enum" : [ "CASH", "CHECK" ]
    },
    "ReceiptStatus" : {
      "type" : "string",
      "enum" : [ "PENDING", "DONE", "CANCEL" ]
    },
    "Receipt" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "trackingNumber" : {
          "type" : "string"
        },
        "sender" : {
          "$ref" : "#/definitions/Sender"
        },
        "recipient" : {
          "$ref" : "#/definitions/Recipient"
        },
        "package" : {
          "$ref" : "#/definitions/Package"
        },
        "declaredValues" : {
          "type" : "string"
        },
        "numberOfPackages" : {
          "type" : "string"
        },
        "insuranceAmount" : {
          "type" : "string"
        },
        "totalWeight" : {
          "type" : "string"
        },
        "unitPrice" : {
          "type" : "number"
        },
        "costPrice" : {
          "type" : "number"
        },
        "extraCost" : {
          "type" : "number"
        },
        "totalAmount" : {
          "type" : "number"
        },
        "totalCostPrice" : {
          "type" : "number"
        },
        "createdDate" : {
          "type" : "string"
        },
        "signedDate" : {
          "type" : "string"
        },
        "senderSignature" : {
          "type" : "boolean"
        },
        "senderId" : {
          "$ref" : "#/definitions/Sender"
        },
        "receiverSignature" : {
          "type" : "boolean"
        }
      }
    }
  }
}